+++ client/app/src/main/java/com/example/witt/data/model/remote/plan/get_plan/GetPlanResponse.kt
@@ -1,6 +1,6 @@
-package com.example.witt.data.model.plan.get_plan
+package com.example.witt.data.model.remote.plan.get_plan
 
-import com.example.witt.domain.model.plan.get_plan.*
+import com.example.witt.domain.model.remote.plan.get_plan.*
 import com.example.witt.utils.convertIsoToDate
 
 data class GetPlanResponse (
