+++ client/app/src/main/java/com/example/witt/data/model/remote/plan/join_plan/JoinPlanResponse.kt
@@ -1,6 +1,6 @@
-package com.example.witt.data.model.plan.join_plan
+package com.example.witt.data.model.remote.plan.join_plan
 
-import com.example.witt.domain.model.plan.join_plan.JoinPlanModel
+import com.example.witt.domain.model.remote.plan.join_plan.JoinPlanModel
 
 data class JoinPlanResponse (
     val status: Boolean,
