+++ client/app/src/main/java/com/example/witt/data/model/remote/plan/make_plan/response/MakePlanResponse.kt
@@ -1,4 +1,4 @@
-package com.example.witt.data.model.plan.make_plan.response
+package com.example.witt.data.model.remote.plan.make_plan.response
 
 data class MakePlanResponse (
     val status: Boolean,
