+++ client/app/src/main/java/com/example/witt/data/source/remote/plan/get_plan/GetPlanDataSource.kt
@@ -1,6 +1,6 @@
 package com.example.witt.data.source.remote.plan.get_plan
 
-import com.example.witt.data.model.plan.get_plan.GetPlanResponse
+import com.example.witt.data.model.remote.plan.get_plan.GetPlanResponse
 
 interface GetPlanDataSource {
     suspend fun getPlan(planId: Int): Result<GetPlanResponse>
