+++ client/app/src/main/java/com/example/witt/data/source/remote/plan/get_plan/GetPlanListDataSource.kt
@@ -1,6 +1,6 @@
 package com.example.witt.data.source.remote.plan.get_plan
 
-import com.example.witt.data.model.plan.get_plan.GetPlanListResponse
+import com.example.witt.data.model.remote.plan.get_plan.GetPlanListResponse
 
 interface GetPlanListDataSource {
     suspend fun getPlanList(): Result<GetPlanListResponse>
