+++ client/app/src/main/java/com/example/witt/data/source/remote/plan/join_plan/OutPlanDataSource.kt
@@ -1,6 +1,6 @@
 package com.example.witt.data.source.remote.plan.join_plan
 
-import com.example.witt.data.model.plan.join_plan.OutPlanResponse
+import com.example.witt.data.model.remote.plan.out_plan.OutPlanResponse
 
 interface OutPlanDataSource {
