+++ client/app/src/main/java/com/example/witt/data/source/remote/plan/join_plan/OutPlanDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package com.example.witt.data.source.remote.plan.join_plan
 
 import com.example.witt.data.api.PlanService
-import com.example.witt.data.model.plan.join_plan.OutPlanResponse
+import com.example.witt.data.model.remote.plan.out_plan.OutPlanResponse
 import com.example.witt.di.IoDispatcher
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
