+++ client/app/src/main/java/com/example/witt/data/source/remote/plan/make_plan/MakePlanDataSource.kt
@@ -1,7 +1,7 @@
 package com.example.witt.data.source.remote.plan.make_plan
 
-import com.example.witt.data.model.plan.make_plan.request.MakePlanRequest
-import com.example.witt.data.model.plan.make_plan.response.MakePlanResponse
+import com.example.witt.data.model.remote.plan.make_plan.request.MakePlanRequest
+import com.example.witt.data.model.remote.plan.make_plan.response.MakePlanResponse
 
 interface MakePlanDataSource {
