+++ client/app/src/main/java/com/example/witt/data/source/remote/plan/make_plan/MakePlanDataSourceImpl.kt
@@ -1,8 +1,8 @@
 package com.example.witt.data.source.remote.plan.make_plan
 
 import com.example.witt.data.api.PlanService
-import com.example.witt.data.model.plan.make_plan.request.MakePlanRequest
-import com.example.witt.data.model.plan.make_plan.response.MakePlanResponse
+import com.example.witt.data.model.remote.plan.make_plan.request.MakePlanRequest
+import com.example.witt.data.model.remote.plan.make_plan.response.MakePlanResponse
 import com.example.witt.di.IoDispatcher
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
