+++ client/app/src/main/java/com/example/witt/data/source/remote/user/GetUserInfoDataSource.kt
@@ -1,6 +1,6 @@
 package com.example.witt.data.source.remote.user
 
-import com.example.witt.data.model.user.response.GetUserInfoResponse
+import com.example.witt.data.model.remote.user.response.GetUserInfoResponse
 
 interface GetUserInfoDataSource {
     suspend fun getUserInfo() : Result<GetUserInfoResponse>
