+++ client/app/src/main/java/com/example/witt/data/source/remote/user/GetUserInfoDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package com.example.witt.data.source.remote.user
 
 import com.example.witt.data.api.UserService
-import com.example.witt.data.model.user.response.GetUserInfoResponse
+import com.example.witt.data.model.remote.user.response.GetUserInfoResponse
 import com.example.witt.di.IoDispatcher
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
