--- client/app/src/main/java/com/example/witt/domain/model/plan/adapter/PlanModel.kt
@@ -1,9 +1,0 @@
-package com.example.witt.domain.model.plan.adapter
-
-import com.example.witt.domain.model.plan.get_plan.DetailPlanModel
-
-data class PlanModel (
-    val day: String,
-    val date: String,
-    val detailPlan: List<DetailPlanModel>
-        )
