--- client/app/src/main/java/com/example/witt/domain/model/plan/remove_plan/RemovePlanModel.kt
@@ -1,6 +1,0 @@
-package com.example.witt.domain.model.plan.remove_plan
-
-data class RemovePlanModel (
-    val status: Boolean,
-    val reason: String
-        )
