+++ client/app/src/main/java/com/example/witt/domain/model/remote/plan/get_plan/GetPlanListModel.kt
@@ -1,6 +1,6 @@
-package com.example.witt.domain.model.plan.get_plan
+package com.example.witt.domain.model.remote.plan.get_plan
 
-import com.example.witt.domain.model.plan.PlanStateModel
+import com.example.witt.domain.model.use_case.plan.PlanStateModel
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
