+++ client/app/src/main/java/com/example/witt/domain/model/remote/plan/join_plan/JoinPlanModel.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.plan.join_plan
+package com.example.witt.domain.model.remote.plan.join_plan
 
 data class JoinPlanModel(
     val status: Boolean,
