+++ client/app/src/main/java/com/example/witt/domain/model/remote/plan/make_plan/MakePlanModel.kt
@@ -1,6 +1,6 @@
-package com.example.witt.data.model.plan.make_plan.request
+package com.example.witt.domain.model.remote.plan.make_plan
 
-data class MakePlanRequest (
+data class MakePlanModel (
     val StartDate: String,
     val EndDate: String,
     val Name: String,
