+++ client/app/src/main/java/com/example/witt/domain/model/remote/plan/make_plan/MakePlanResponseModel.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.plan.make_plan
+package com.example.witt.domain.model.remote.plan.make_plan
 
 class MakePlanResponseModel (
     val status: Boolean,
