+++ client/app/src/main/java/com/example/witt/domain/model/remote/user/GetUserInfoModel.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.user
+package com.example.witt.domain.model.remote.user
 
 data class GetUserInfoModel (
     val status: Boolean,
