+++ client/app/src/main/java/com/example/witt/domain/repository/PlanRepository.kt
@@ -1,12 +1,11 @@
 package com.example.witt.domain.repository
 
-import com.example.witt.domain.model.plan.get_plan.GetPlanListModel
-import com.example.witt.domain.model.plan.get_plan.GetPlanModel
-import com.example.witt.domain.model.plan.make_plan.MakePlanModel
-import com.example.witt.domain.model.plan.make_plan.MakePlanResponseModel
-import com.example.witt.domain.model.plan.join_plan.JoinPlanModel
-import com.example.witt.domain.model.plan.join_plan.OutPlanModel
-import com.example.witt.domain.model.plan.remove_plan.RemovePlanModel
+import com.example.witt.domain.model.remote.plan.get_plan.GetPlanListModel
+import com.example.witt.domain.model.remote.plan.get_plan.GetPlanModel
+import com.example.witt.domain.model.remote.plan.make_plan.MakePlanModel
+import com.example.witt.domain.model.remote.plan.make_plan.MakePlanResponseModel
+import com.example.witt.domain.model.remote.plan.join_plan.JoinPlanModel
+import com.example.witt.domain.model.remote.plan.out_plan.OutPlanModel
 
 interface PlanRepository {
