+++ client/app/src/main/java/com/example/witt/domain/use_case/remote/MakePlanUseCase.kt
@@ -1,7 +1,7 @@
 package com.example.witt.domain.use_case.remote
 
-import com.example.witt.domain.model.plan.make_plan.MakePlanModel
-import com.example.witt.domain.model.plan.make_plan.MakePlanResponseModel
+import com.example.witt.domain.model.remote.plan.make_plan.MakePlanModel
+import com.example.witt.domain.model.remote.plan.make_plan.MakePlanResponseModel
 import com.example.witt.domain.repository.PlanRepository
 import javax.inject.Inject
